moved bohs_net_detector.py to utils directory
